Move core/text/lingo -> test/robot/lingo.
