+++ test/robot/lingo/examples/CMakeBuild.cmake
