+++ test/robot/lingo/examples/calc/.gitignore
