+++ test/robot/lingo/examples/calc/CMakeBuild.cmake
