+++ test/robot/lingo/examples/cstcalc/.gitignore
