+++ test/robot/lingo/examples/cstcalc/CMakeBuild.cmake
