+++ test/robot/lingo/examples/simplecalc/CMakeBuild.cmake
