+++ test/robot/lingo/examples/simplecalc/main.lingo
@@ -25,7 +25,7 @@ import (
 
 	"github.com/google/gapid/core/app"
 	"github.com/google/gapid/core/log"
-	"github.com/google/gapid/core/text/lingo"
+	"github.com/google/gapid/test/robot/lingo"
 )
 
 func main() {
@@ -40,8 +40,6 @@ type (
 	Digits     string
 	Operator   string
 	Bracket    string
-
-	Structural struct{}
 )
 
 const (
@@ -66,114 +64,68 @@ func run(ctx context.Context) error {
 		app.Usage(ctx, "Missing expression")
 		return nil
 	}
-	r := lingo.Records{}
-	s := lingo.NewStringScanner(ctx, "command_line", strings.Join(args, " "), &r)
-	_, err := expression(s)
+	input := strings.Join(args, " ")
+	s := lingo.NewStringScanner(ctx, "command_line", input, nil)
+	value, err := expression(s)
 	if err != nil {
 		return err
 	}
 	if !s.EOF() {
 		return log.Err(ctx, nil, "Input not consumed")
 	}
-	cst := r.ToCST()
-	result := eval(ctx, cst)
-	fmt.Fprintf(os.Stderr, "%v = %d\n", cst, result.value)
+	fmt.Fprintf(os.Stdout, "%s = %d\n", input, value)
 	return nil
 }
 
-type Eval struct {
-	hasValue bool
-	value    int
-	op       Operator
+func expression(s *lingo.Scanner) (int, error) {
+	return addition(s)
 }
 
-func eval(ctx context.Context, node *lingo.Node) Eval {
-	result := Eval{}
-	switch obj := node.Object.(type) {
-	case int:
-		result.hasValue = true
-		result.value = obj
-	case Operator:
-		result.op = obj
-	}
-	for _, child := range node.Children {
-		got := eval(ctx, child)
-		if got.hasValue {
-			if !result.hasValue {
-				result.value = got.value
-			} else {
-				switch result.op {
-				case opAdd:
-					result.value += got.value
-				case opSubtract:
-					result.value -= got.value
-				case opMultiply:
-					result.value *= got.value
-				case opDivide:
-					result.value /= got.value
-				default:
-					log.F(ctx, "Cannot combine values without operator")
-				}
-			}
-			result.hasValue = true
-		}
-		if got.op != "" {
-			result.op = got.op
-		}
-	}
-	return result
-}
-
-func expression(s *lingo.Scanner) (Structural, error) {
-	addition(s)
-	return Structural{}, nil
-}
-
-func parenthesised(s *lingo.Scanner) (Structural, error) {
+func parenthesised(s *lingo.Scanner) (int, error) {
 	openParenthesis(s)
-	expression(s)
+	value := expression(s)
 	closeParenthesis(s)
-	return Structural{}, nil
+	return value, nil
 }
 
-func addition(s *lingo.Scanner) (Structural, error) {
-	multiplication(s)
+func addition(s *lingo.Scanner) (int, error) {
+	value := multiplication(s)
 	for {
 		space(s)
 		switch {
 		case opAdd(s):
-			multiplication(s)
+			value += multiplication(s)
 		case opSubtract(s):
-			multiplication(s)
+			value -= multiplication(s)
 		default:
-			return Structural{}, nil
+			return value, nil
 		}
 	}
 }
 
-func multiplication(s *lingo.Scanner) (Structural, error) {
-	numeral(s)
+func multiplication(s *lingo.Scanner) (int, error) {
+	value := numeral(s)
 	for {
 		space(s)
 		switch {
 		case opMultiply(s):
-			numeral(s)
+			value *= numeral(s)
 		case opDivide(s):
-			numeral(s)
+			value /= numeral(s)
 		default:
-			return Structural{}, nil
+			return value, nil
 		}
 	}
 }
 
-func numeral(s *lingo.Scanner) (Structural, error) {
-	switch {
-	case parenthesised(s):
-	case number(s):
-	default:
-		return Structural{}, s.Error(nil, ErrorParenOrNumber)
+func numeral(s *lingo.Scanner) (int, error) {
+	if value, err := parenthesised(s); err == nil {
+		return value, err
 	}
-	return Structural{}, nil
+	if value, err := number(s); err == nil {
+		return value, err
+	}
+	return 0, s.Error(nil, ErrorParenOrNumber)
 }
 
 func number(s *lingo.Scanner) (int, error) {
