+++ test/robot/lingo/generator/introspect.go
