+++ test/robot/lingo/generator/rewrite.go
