+++ test/robot/lingo/scanner.go
