+++ test/robot/search/script/constants.lingo
@@ -17,7 +17,7 @@ package script
 import (
 	"regexp"
 
-	"github.com/google/gapid/core/text/lingo"
+	"github.com/google/gapid/test/robot/lingo"
 )
 
 type _ lingo.Scanner
