+++ test/robot/search/script/doc.go
@@ -24,6 +24,6 @@ package script
 // Having these here helps out tools that can't cope with missing dependancies
 import (
 	_ "github.com/google/gapid/core/log"
-	_ "github.com/google/gapid/core/text/lingo"
+	_ "github.com/google/gapid/test/robot/lingo"
 	_ "github.com/google/gapid/test/robot/search/query"
 )
