+++ test/robot/search/script/expression.lingo
@@ -15,7 +15,7 @@
 package script
 
 import (
-	"github.com/google/gapid/core/text/lingo"
+	"github.com/google/gapid/test/robot/lingo"
 	"github.com/google/gapid/test/robot/search/query"
 )
