+++ test/robot/search/script/literal.lingo
@@ -17,7 +17,7 @@ package script
 import (
 	"strconv"
 
-	"github.com/google/gapid/core/text/lingo"
+	"github.com/google/gapid/test/robot/lingo"
 	"github.com/google/gapid/test/robot/search/query"
 )
